backend/requirements.txt
fastapi
uvicorn
passlib
bcrypt
python-jose
requests
pydantic
alembic
fastapi-sqlalchemy